Comments around Node required code were incorrect. Use of JS files broke the gulp "pull-code" task. Should probably rename all GS file as JS
