+++ src/jiraIssueStatusUpdates/issueTransitioner.gs
@@ -1,9 +1,13 @@
-// Require imports
+// Node required code block
 const Request = require('../jiraApi.gs');
 const getIssue = require('../jiraCommon.gs').getIssue;
-const IssueTransitionConfiguration = require('./issueTransitionConfiguration.js');
-// End of Require imports
+const IssueTransitionConfiguration = require('./issueTransitionConfiguration.gs');
+// End of Node required code block
 
+/**
+ * Gets the list of transitions valid for the specified issue in its current status
+ * @param {string} issueKey the id of the issue
+ */
 function getPossibleTransitions(issueKey) {
   var response = {
     'data': [],
@@ -42,8 +46,12 @@ function getPossibleTransitions(issueKey) {
   return response;
 }
 
+/**
+ * Makes a request to the JIRA API to apply the transition to an issue
+ * @param {string} issueKey issue key 
+ * @param {integer} transitionId the id of the transition to move it to the desired state
+ */
 function makeTransition(issueKey, transitionId) {
-
   var request = new Request();
   request.call('issueTransitionUpdate', {issueIdOrKey: issueKey,"transition": {"id": transitionId}});
   var resp = request.getResponse();
@@ -58,8 +66,18 @@ function makeTransition(issueKey, transitionId) {
 
 }
 
+/**
+ * Constuctor for transitioning an issue from one status to another
+ * checks current status of an issue first
+ * gets the transitions possible from moving from that status
+ * applies (if possible) the new status to the issue
+ */
 function IssueTransitioner() {
     var config = new IssueTransitionConfiguration();
+    /**
+     * @param issueKey the key of the issue to transition
+     * @param newStatus the status that it is desired to transition to.
+     */
     this.transition = function(issueKey,newStatus) {
         newStatus = newStatus.toLowerCase();
         var result = getIssue(issueKey);
@@ -98,4 +116,6 @@ function IssueTransitioner() {
     }
 }
 
+// Node required code block
 module.exports = IssueTransitioner;
+// End of Node required code block
